// packages/protocol/src/generated/codex/ServerNotification.ts
export type ServerNotification =
  | { method: "error"; params: ErrorNotification }
  | { method: "thread/started"; params: ThreadStartedNotification }
  | { method: "thread/name/updated"; params: ThreadNameUpdatedNotification }
  | {
      method: "thread/tokenUsage/updated"
      params: ThreadTokenUsageUpdatedNotification
    }
  | { method: "turn/started"; params: TurnStartedNotification }
  | { method: "turn/completed"; params: TurnCompletedNotification }
  | { method: "turn/diff/updated"; params: TurnDiffUpdatedNotification }
  | { method: "turn/plan/updated"; params: TurnPlanUpdatedNotification }
  | { method: "item/started"; params: ItemStartedNotification }
  | { method: "item/completed"; params: ItemCompletedNotification }
  | {
      method: "rawResponseItem/completed"
      params: RawResponseItemCompletedNotification
    }
  | { method: "item/agentMessage/delta"; params: AgentMessageDeltaNotification }
  | { method: "item/plan/delta"; params: PlanDeltaNotification }
  | {
      method: "item/commandExecution/outputDelta"
      params: CommandExecutionOutputDeltaNotification
    }
  | {
      method: "item/commandExecution/terminalInteraction"
      params: TerminalInteractionNotification
    }
  | {
      method: "item/fileChange/outputDelta"
      params: FileChangeOutputDeltaNotification
    }
  | {
      method: "item/mcpToolCall/progress"
      params: McpToolCallProgressNotification
    }
  | {
      method: "mcpServer/oauthLogin/completed"
      params: McpServerOauthLoginCompletedNotification
    }
  | { method: "account/updated"; params: AccountUpdatedNotification }
  | {
      method: "account/rateLimits/updated"
      params: AccountRateLimitsUpdatedNotification
    }
  | { method: "app/list/updated"; params: AppListUpdatedNotification }
  | {
      method: "item/reasoning/summaryTextDelta"
      params: ReasoningSummaryTextDeltaNotification
    }
  | {
      method: "item/reasoning/summaryPartAdded"
      params: ReasoningSummaryPartAddedNotification
    }
  | {
      method: "item/reasoning/textDelta"
      params: ReasoningTextDeltaNotification
    }
  | { method: "thread/compacted"; params: ContextCompactedNotification }
  | { method: "deprecationNotice"; params: DeprecationNoticeNotification }
  | { method: "configWarning"; params: ConfigWarningNotification }
  | {
      method: "windows/worldWritableWarning"
      params: WindowsWorldWritableWarningNotification
    }
  | {
      method: "account/login/completed"
      params: AccountLoginCompletedNotification
    }
  | { method: "authStatusChange"; params: AuthStatusChangeNotification }
  | { method: "loginChatGptComplete"; params: LoginChatGptCompleteNotification }
  | { method: "sessionConfigured"; params: SessionConfiguredNotification }